Extract analytics sink to drop web/native logger duplication
WebLogging and NativeLogging shared most of their code: Sentry init
and error/breadcrumb plumbing, param building (detail-flatten +
product injection), console fallthrough — all identical. The only
genuine differences were the SDK calls (gtag vs FirebaseAnalytics),
the consent-gate semantics, and the platform-specific navigate
behaviour.

Split that seam: Logger owns the cross-cutting work; AnalyticsSink
owns whatever the platform's SDK + consent flow demands. WebSink
and NativeSink replace the two old classes.
